backend/requirements.txt
werkzeug
sqlalchemy
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
email-validator==2.3.0
python-dotenv==1.0.1
PyMySQL==1.1.0
